client/requirements.txt
textual-dev==1.7.0
httpx==0.28.1

